Refactor/server controllers (#44)
* chore(tsconfig): rename @controllers/* path to @controllers

* refactor: rename import path from @controllers/[...] to @controllers

* feat: add index.ts to controllers folder

* feat: add index.ts to controllers/filters folder

* refactor: rename import paths from "./filters/[...] to ./filters

replace to work with index.ts file
