+++ server/src/controllers/index.ts
@@ -1,0 +1,16 @@
+export * from "./affixesController";
+export * from "./authController";
+export * from "./chatCommandsController";
+export * from "./configsController";
+export * from "./filesController";
+export * from "./messageCategoriesController";
+export * from "./messagesController";
+export * from "./moodsController";
+export * from "./overlaysController";
+export * from "./redemptionsController";
+export * from "./streamSessionsController";
+export * from "./tagsController";
+export * from "./timersController";
+export * from "./triggersController";
+export * from "./usersController";
+export * from "./widgetsController";
