+++ server/src/controllers/redemptionsController.ts
@@ -1,6 +1,6 @@
 import { NextFunction, Request, Response } from "express";
 import { RequestRedemptionQuery } from "@types";
-import { filterRedemptionsByUrlParams } from "./filters/redemptionsFilter";
+import { filterRedemptionsByUrlParams } from "./filters";
 import { getRedemptions, getRedemptionsCount } from "@services";
 
 export const getRedemptionsList = async (
